// app_taskflow_backend_01/src/types/task.types.ts
import { Task, TaskStatus } from '@prisma/client';

export type TaskPublic = Task;

export interface CreateTaskDto {
    title: string;
    description?: string;
    status?: TaskStatus;
    projectId: string;
    assignedTo?: string;
}
export interface UpdateTaskDto {
    title?: string;
    description?: string;
    status?: TaskStatus;
    assignedTo?: string | null;
}